+++ src/ui/stats.js
@@ -5,7 +5,7 @@
 
 import {PACKAGE_ID, IS_UNITTEST} from '../consts.js';
 import {LibWrapperNotifications} from './notifications.js';
-import {game_user_can} from '../utils/polyfill.js'
+import {game_user_can} from '../shared/polyfill.js'
 
 export class LibWrapperStats {
 	static _collect_stats() {